refactor: from itemsPage to Tools
