+++ src/pages/items-page/ItemsPage.tsx
@@ -70,55 +70,12 @@ export const ItemsPage = (props: IItemsPageProps): JSX.Element => {
     const modifyItems: ModifyItemsType = toolsModel.modifyItemsByParams(catalogItems, toolsSetting);
     const [prods, setProds] = useState(modifyItems.items);
 
-    const setItemsData = () => {
-        const modifyData = toolsModel.modifyItemsByParams(catalogItems, toolsSetting);
-        setProds(modifyData.items);
-        setSearchParams(modifyData.urlParams);
-    }
-
-    const filterItemsOnChange = (e: React.ChangeEvent) => {
-        const elem = e.target as HTMLSelectElement;
-        const value = elem.value;
-        const itemObjectKey = elem.id as keyof FilterType;
-        toolsSetting.filter[itemObjectKey] = value;
-        setItemsData();
-    }
-
-    const sortItemsOnClick = (e: React.MouseEvent) => {
-        toolsSetting.sort.price = null;
-        toolsSetting.sort.store = null;
-        const elem = e.target as HTMLElement;
-        const direction = elem.dataset.direction as 'assent' | 'descent' | null;
-        const option = elem.dataset.option as keyof SortType;
-        toolsSetting.sort[option] = direction;
-        setItemsData();
-    }
-
-    const itemsList: JSX.Element[] = prods.map(elem =>
+    const itemsList: JSX.Element[] = prods.map((elem) =>
         <ItemCard
             item={elem}
             state={props.state}
-            key={elem.id} />);
-
-    const categoriesFilter: JSX.Element[] = categories.map((name) =>
-        <option
-            key={name}
-            value={name}>
-            {name}
-        </option>
-    );
-
-    const brandFilter: JSX.Element[] = brandsInCategory.map((name, i) => {
-        return (
-            <option
-                key={name}
-                value={name}>
-                {name}
-            </option>)
-    });
-
-    let categorySelectValue = toolsSetting.filter.category === null ? '...' : toolsSetting.filter.category;
-    let brandSelectValue = toolsSetting.filter.brand === null ? '...' : toolsSetting.filter.brand;
+            key={elem.id}
+        />);
 
     const setItemsFromTools = (items: IItem[], urlParam: ParamKeyValuePair[]) => {
         setProds(items);
@@ -143,74 +100,9 @@ export const ItemsPage = (props: IItemsPageProps): JSX.Element => {
                     Items in page:
                     <span>{prods.length}</span>
                 </h4>
+
                 <Tools items={[...catalog.products]} setItems={setItemsFromTools} toolsSetting={toolsSetting} />
-                <section className='catalog__wrap__tools-wrap'>
-                    <div className='catalog__wrap__tools-wrap__sort'>
-                        <div className='catalog__wrap__tools-wrap__sort__by-price'>
-                            <span>Sort by price</span>
-                            <div className='catalog__wrap__tools-wrap__sort__by-price__button-wrap'>
-                                <button
-                                    data-direction='assent'
-                                    data-option='price'
-                                    onClick={sortItemsOnClick}
-                                >🠕</button>
-                                <button
-                                    data-direction='descent'
-                                    data-option='price'
-                                    onClick={sortItemsOnClick}
-                                >🠗</button>
-                            </div>
 
-                        </div>
-                        <div className='catalog__wrap__tools-wrap__sort__by-stock'>
-                            <span>Sort by stock</span>
-                            <div className='catalog__wrap__tools-wrap__sort__by-stock__button-wrap'>
-                                <button
-                                    data-direction='assent'
-                                    data-option='stock'
-                                    onClick={sortItemsOnClick}
-                                >🠕</button>
-                                <button
-                                    data-direction='descent'
-                                    data-option='stock'
-                                    onClick={sortItemsOnClick}
-                                >🠗</button>
-                            </div>
-                        </div>
-                    </div>
-                    <div className='catalog__wrap__tools-wrap__filter'>
-                        <div className='catalog__wrap__tools-wrap__filter__wrap'>
-                            <h3>Category</h3>
-                            <select
-                                className='catalog__wrap__tools-wrap__filter__wrap__category'
-                                id='category'
-                                onChange={filterItemsOnChange}
-                                value={categorySelectValue}>
-                                <option value='...'>...</option>
-                                {categoriesFilter}
-                            </select>
-                        </div>
-                        <div className='catalog__wrap__tools-wrap__filter__wrap'>
-                            <h3>Brand</h3>
-                            <select
-                                id='brand'
-                                className='catalog__wrap__tools-wrap__filter__wrap__brand'
-                                onChange={filterItemsOnChange}
-                                value={brandSelectValue}>
-                                <option value='...'>...</option>
-                                {brandFilter}
-                            </select>
-                        </div>
-
-                    </div>
-                    <div className='catalog__wrap__tools-wrap__search'>
-                        <input
-                            className='catalog__wrap__tools-wrap__search__input'
-                            type='text'
-                            placeholder="Product search" />
-                        <button className='blue-button'>Search</button>
-                    </div>
-                </section>
                 <section className='catalog__wrap__items'>
                     <h3 className='catalog__wrap__items__title'>
                         Products
@@ -220,9 +112,7 @@ export const ItemsPage = (props: IItemsPageProps): JSX.Element => {
                         {itemsList}
                     </div>
                 </section>
-
             </div >
         </section >
     )
 }
-
